backend/requirements.txt
asgiref==3.5.2
Django==2.2.19
django-colorfield==0.7.1
django-cors-headers==3.13.0
django-filter==2.2.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
django-extra-fields==3.0.2
gunicorn==20.1.0
pytz==2021.1
psycopg2-binary==2.8.6
Pillow==9.2.0
PyJWT==2.4.0
pytz==2021.1
sqlparse==0.4.2
fpdf==1.7.2
